добавлен blueprint user, скорректированы шаблоны (#18)
