+++ webapp/__init__.py
@@ -1,9 +1,10 @@
-from flask import Flask, render_template, flash, redirect, url_for
-from flask_login import LoginManager, current_user, login_required, login_user, logout_user
+from flask import Flask, render_template
+from flask_login import LoginManager, current_user, login_required
 
-from webapp.forms import LoginForm
-from webapp.model import db, User
 
+from webapp.model import db
+from webapp.user.models import User
+from webapp.user.views import blueprint as user_blueprint
 
 def create_app():
     app = Flask(__name__)
@@ -12,7 +13,8 @@ def create_app():
 
     login_manager = LoginManager()
     login_manager.init_app(app)
-    login_manager.login_view = 'login'
+    login_manager.login_view = 'user.login'
+    app.register_blueprint(user_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
@@ -23,37 +25,12 @@ def create_app():
         title = "Главная страница"
         return render_template('index.html', page_title=title)
     
-    @app.route('/login')
-    def login():
-        if current_user.is_authenticated:
-            return redirect(url_for('index'))
-        title = 'Авторизация'
-        login_form = LoginForm()
-        return render_template('login.html', page_title=title, form=login_form)
-
+    
     @app.route('/about')
     def about():
         title = "О нас"
         return render_template('about.html', page_title=title)
     
-    @app.route('/process-login', methods=['POST'])
-    def process_login():
-        form = LoginForm()
-
-        if form.validate_on_submit():
-            user = User.query.filter(User.username == form.username.data).first()
-            if user and user.check_password(form.password.data):
-                login_user(user, remember=form.remember_me.data)
-                flash('Вы успешно вошли на сайт')
-                return redirect(url_for('login'))
-        flash('Неправильное имя пользователя или пароль')
-        return redirect(url_for('login'))
-    
-    @app.route('/logout')
-    def logout():
-        logout_user()
-        flash('Вы вышли из личного кабинета')
-        return redirect(url_for('index'))
 
     @app.route('/admin')
     @login_required
