+++ webapp/templates/base.html
@@ -15,10 +15,10 @@
       <!-- menu -->
       <ul class="nav nav-pills">
         <li class="nav-item"><a href="/" class="nav-link active" aria-current="page">Home</a></li>
-        <li class="nav-item"><a href="/login" class="nav-link">Авторизоваться</a></li>
+        <li class="nav-item"><a href="/users/login" class="nav-link">Авторизоваться</a></li>
         <li class="nav-item"><a href="/about" class="nav-link">О нас</a></li>
         <li class="nav-item"><a href="#" class="nav-link">FAQs</a></li>
-        <li class="nav-item"><a href="#" class="nav-link">Контакты</a></li>
+        <li class="nav-item"><a href="/users/logout" class="nav-link">logout</a></li>
       </ul>
      </header>
      </div>
